Moves app class to top-level package to fix keydown bug
