+++ MobileSyncExplorerKotlinTemplate/app/AndroidManifest.xml
@@ -3,7 +3,7 @@
     package="com.salesforce.mobilesyncexplorerkotlintemplate">
 
     <application
-        android:name="com.salesforce.mobilesyncexplorerkotlintemplate.app.MobileSyncExplorerKotlinTemplateApp"
+        android:name="com.salesforce.mobilesyncexplorerkotlintemplate.MobileSyncExplorerKotlinTemplateApp"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
