+++ MobileSyncExplorerKotlinTemplate/app/src/main/java/com/salesforce/mobilesyncexplorerkotlintemplate/MobileSyncExplorerKotlinTemplateApp.kt
@@ -24,7 +24,7 @@
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  */
-package com.salesforce.mobilesyncexplorerkotlintemplate.app
+package com.salesforce.mobilesyncexplorerkotlintemplate
 
 import android.app.Application
 import android.content.Context
