+++ MobileSyncExplorerKotlinTemplate/app/src/main/java/com/salesforce/mobilesyncexplorerkotlintemplate/core/repos/SObjectSyncableRepoBase.kt
@@ -99,8 +99,7 @@ abstract class SObjectSyncableRepoBase<T : SObject>(
         syncMutex.withLockDebug {
             doSyncDown()
             try {
-                // TODO This cannot be called immediately after syncDown impl with coroutines
-                //  UPDATE Wolf just opened https://github.com/forcedotcom/SalesforceMobileSDK-Android/pull/2301 which would fix the issue
+                // TODO when is the appropriate time to call clean resync ghosts?
                 syncManager.suspendCleanResyncGhosts(syncName = syncDownName)
             } catch (ex: CleanResyncGhostsException) {
                 throw SyncDownException.CleaningUpstreamRecordsFailed(cause = ex)
